fixed publications page. moved shared form components to 'components' folder
