+++ src/client/src/routes/IndexPage/components/ProjectsDetails.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Input from './elements/Input';
+import Input from '../../../components/form/Input';
 
 const DATA_KEY = 'currentProjects';
