+++ src/server/data/publications-data.json
@@ -1,1 +1,46 @@
-[{"name":"654","description":"321","year":"32","authors":"15","topic":"4","type":"9","publisher":"84","publicationStatus":"654","fileLinks":[{"info":"12","link":"5","type":"666"}]}]
+[
+  {
+    "name": "321",
+    "description": "321",
+    "year": "321",
+    "authors": "",
+    "topic": "51",
+    "type": "546",
+    "publisher": "54",
+    "publicationStatus": "7474",
+    "fileLinks": [
+      {
+        "info": "151",
+        "type": "2",
+        "link": "3"
+      },
+      {
+        "info": "45",
+        "link": "6"
+      },
+      {
+        "type": "5"
+      },
+      {
+        "type": "00"
+      }
+    ]
+  },
+  {
+    "name": "sd",
+    "description": "k",
+    "year": "km",
+    "authors": "",
+    "topic": "m,",
+    "type": "m,",
+    "publisher": "mkmk",
+    "publicationStatus": "kjlkj",
+    "fileLinks": [
+      {
+        "info": "m.",
+        "type": ",m.,m.,m.,m",
+        "link": "m,,m,m"
+      }
+    ]
+  }
+]
